requirements-dev.txt
-r requirements-base.txt
pandas==2.2.3
scipy==1.15.3
seaborn==0.13.2
tqdm==4.67.1
matplotlib==3.10.3
roboflow==1.1.65
thop==0.1.1.post2209072238
networkx==3.5
contourpy==1.3.2
cycler==0.12.1
mpmath==1.3.0
pytest==8.0.2
pytest-django==4.8.0
black==24.2.0
flake8==7.0.0 